Vibe/requirements.txt
AppKit==0.2.8
gymnasium==1.2.1
mss==10.1.0
numpy==2.3.3
opencv_python==4.12.0.88
Pillow==11.3.0
pyautogui==0.9.54
pyobjc_framework_cocoa==11.1
pyobjc_framework_quartz==11.1
pytesseract==0.3.13
Quartz==0.0.1.dev0
stable_baselines3==2.7.0
